### 2022-10-27 #109 @whedon
[PRE REVIEW]: DelayDiffEq: Generating Delay Differential Equation Solvers via Recursive Embedding of Ordinary Differential Equation Solvers
**Submitting author:** @devmotion (<a href="http://orcid.org/0000-0001-9282-053X">David Widmann</a>)
**Repository:** <a href="https://github.com/SciML/DelayDiffEq.jl" target ="_blank">https://github.com/SciML/DelayDiffEq.jl</a>
**Version:**
