### 2022-10-31 2d48582
`nW` should be an upper bound of `nJ`
Ref: https://github.com/SciML/OrdinaryDiffEq.jl/pull/1780
