### 2022-11-01 #109 @devmotion
Strange, it worked locally (and with Github CI) by following the instructions of the template. Possibly a different ruby version? I'll try to fix it in the next few days.
