### 2022-11-09 d682761
update symbol return codes to ReturnCode
