+++ test/integrators/retcode.jl
@@ -8,11 +8,11 @@ const prob = prob_dde_constant_1delay_ip
                         constrained = false)
 
     sol1 = solve(prob, alg)
-    @test sol1.retcode == :Success
+    @test sol1.retcode == ReturnCode.Success
 
     sol2 = solve(prob, alg; maxiters = 1, verbose = false)
-    @test sol2.retcode == :MaxIters
+    @test sol2.retcode == ReturnCode.MaxIters
 
     sol3 = solve(prob, alg; dtmin = 5, verbose = false)
-    @test sol3.retcode == :DtLessThanMin
+    @test sol3.retcode == ReturnCode.DtLessThanMin
 end
