### 2022-11-10 49673be
update symbol return codes to ReturnCode (#248)
