+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DelayDiffEq"
 uuid = "bcd4f6db-9728-5f36-b5f7-82caef46ccdb"
 authors = ["Chris Rackauckas <accounts@chrisrackauckas.com>"]
-version = "5.38.3"
+version = "5.39.0"
 
 [deps]
 ArrayInterface = "4fba245c-0d91-5ea0-9b3e-6abc04ee57a9"
@@ -25,6 +25,6 @@ NonlinearSolve = "0.3"
 OrdinaryDiffEq = "6.19.3"
 RecursiveArrayTools = "2"
 Reexport = "0.2, 1.0"
-SciMLBase = "1.34"
+SciMLBase = "1.70"
 UnPack = "0.1, 1.0"
 julia = "1.6"
