+++ paper/paper.yml
@@ -9,7 +9,7 @@ authors:
     affiliation: 1
   - name: Chris Rackauckas\texorpdfstring{~\orcidlink{0000-0001-5850-0663}}{}
     orcid: 0000-0001-5850-0663
-    affiliation: 2,3,4
+    affiliation: "2,3,4"
 affiliations:
   - name: Uppsala University, Sweden
     index: 1
