### 2022-11-18 #109 @devmotion
Still haven't managed to reproduce this locally, is there really no simple way to run the build pipeline locally? Debugging with CI seems a bit suboptimal...

However, I found https://github.com/openjournals/whedon/blob/92346a0773a4003bf0ef361b661dc2787f492936/lib/whedon/author.rb#L69-L70 and https://bugs.ruby-lang.org/issues/12451, so maybe that's the underlying issue. I tried to address that, let's see if it can be compiled now.
