### 2022-11-18 #109 @devmotion
I found https://github.com/JuliaCon/proceedings-review/issues/79#issuecomment-984731999 - should mathtools be added to the list of available LaTeX packages, similar to https://github.com/openjournals/heroku-buildpack-tex/pull/2, or should it be added to the repo? I'm also not really sure what would have to be added to the repo - it seems there is not mathtools.sty in https://www.ctan.org/tex-archive/macros/latex/contrib/mathtools?
