### 2022-11-24 3f1ebba
Move to SimpleNonlinearSolve (#250)
* Move to SimpleNonlinearSolve

* Format

* Interval
