### 2022-11-24 a1db0ac
Fix tests for removal of ODECompositeSolution
