+++ Project.toml
@@ -21,7 +21,7 @@ UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 ArrayInterface = "3, 4, 5, 6"
 DataStructures = "0.18"
 DiffEqBase = "6"
-OrdinaryDiffEq = "6.37"
+OrdinaryDiffEq = "6.33"
 RecursiveArrayTools = "2"
 Reexport = "0.2, 1.0"
 SciMLBase = "1.70"
