### 2022-11-25 51697da
Fix tests for removal of ODECompositeSolution (#251)
* Fix tests for removal of ODECompositeSolution

* Update Project.toml

* fix tests
