+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DelayDiffEq"
 uuid = "bcd4f6db-9728-5f36-b5f7-82caef46ccdb"
 authors = ["Chris Rackauckas <accounts@chrisrackauckas.com>"]
-version = "5.40.0"
+version = "5.40.1"
 
 [deps]
 ArrayInterface = "4fba245c-0d91-5ea0-9b3e-6abc04ee57a9"
@@ -21,7 +21,7 @@ UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 ArrayInterface = "3, 4, 5, 6"
 DataStructures = "0.18"
 DiffEqBase = "6"
-OrdinaryDiffEq = "6.19.3"
+OrdinaryDiffEq = "6.33"
 RecursiveArrayTools = "2"
 Reexport = "0.2, 1.0"
 SciMLBase = "1.70"
