### 2022-11-30 6f1c362
Don't depend on DiffEqBase nlsolver (#252)
* Don't depend on DiffEqBase nlsolver

* format
