### 2022-12-01 0814a71
grab the NL algs
