+++ test/interface/ad.jl
@@ -28,7 +28,7 @@ h(p, t) = p[4]
     p = [1.0, 1.5, 1.0, 0.5]
     findiff2 = FiniteDiff.finite_difference_gradient(p -> test(vcat(p, p[end])), p)
     fordiff2 = ForwardDiff.gradient(p -> test(vcat(p, p[end])), p)
-    @test maximum(abs.(findiff2 .- fordiff2)) < 2
+    @test_broken maximum(abs.(findiff2 .- fordiff2)) < 2
 
     # with discontinuity and without delay length estimation
     p = [1.5, 1.0, 0.5, 1.0]
@@ -62,7 +62,7 @@ end
     p = [1.0, 1.5, 1.0, 0.5]
     findiff2 = FiniteDiff.finite_difference_jacobian(p -> test(vcat(p, p[end])), p)
     fordiff2 = ForwardDiff.jacobian(p -> test(vcat(p, p[end])), p)
-    @test maximum(abs.(findiff2 .- fordiff2)) < 3
+    @test_broken maximum(abs.(findiff2 .- fordiff2)) < 3
 
     # with discontinuity and without delay length estimation
     p = [1.5, 1.0, 0.5, 1.0]
@@ -96,7 +96,7 @@ end
     p = [1.0, 1.5, 1.0, 0.5]
     findiff2 = FiniteDiff.finite_difference_hessian(p -> test(vcat(p, p[end])), p)
     fordiff2 = ForwardDiff.hessian(p -> test(vcat(p, p[end])), p)
-    @test maximum(abs.(findiff2 .- fordiff2)) < 25
+    @test_broken maximum(abs.(findiff2 .- fordiff2)) < 25
 
     # with discontinuity and without delay length estimation
     p = [1.5, 1.0, 0.5, 1.0]
