### 2022-12-12 #109 @odow
The easiest solution is just to remove the dependence on `mathtools`. Why is it needed?
