+++ paper/header.tex
@@ -3,7 +3,7 @@
 \title{DelayDiffEq: Generating Delay Differential Equation Solvers via Recursive Embedding of Ordinary Differential Equation Solvers}
 
 \author[1]{David Widmann\texorpdfstring{~\orcidlink{0000-0001-9282-053X}}{}}
-\author[234]{Chris Rackauckas\texorpdfstring{~\orcidlink{0000-0001-5850-0663}}{}}
+\author[2,3,4]{Chris Rackauckas\texorpdfstring{~\orcidlink{0000-0001-5850-0663}}{}}
 \affil[1]{Uppsala University, Sweden}
 \affil[2]{Massachusetts Institute of Technology, USA}
 \affil[3]{Julia Computing}
