### 2022-12-12 #109 @devmotion
I always use/load `mathtools` instead of `amsmath` since it fixes some bugs in `amsmath` and provides useful additional functionality (see, e.g., https://ftp.acc.umu.se/mirror/CTAN/macros/latex/contrib/mathtools/mathtools.pdf). Here, it seems it is only required for `\coloneqq` commands. I replaced them with `:=` which hopefully should fix whedon's compilation issues. I think it would be useful nevertheless to support `mathtools`, as I said it's *the* package I load for math support in all my LaTeX files and at least I assumed that it is a quite common package.
