### 2022-12-12 #109 @devmotion
Well well, these LaTeX issues are a bit annoying - my feeling is that this could be improved a bit :smile: Even more so since the paper compiles just fine with standard TeXlive installations (as it was the case for the initial submission and all updates) and at least to me it is still unclear what exactly is supported and why (not).

I guess the easiest solution to the latest error is to include the orcidlink.sty file.
