### 2022-12-12 #109 @odow
> Well well, these LaTeX issues are a bit annoying - my feeling is that this could be improved a bit 😄 Even more so since the paper compiles just fine with standard TeXlive installations (as it was the case for the initial submission and all updates) and at least to me it is still unclear what exactly is supported and why (not).
>
> I guess the easiest solution to the latest error is to include the orcidlink.sty file

Just use the minimal set of packages. Do you really need all the extras?

The bot uses a Docker image with a fixed set of packages. There's essentially zero free time on the editorial team to improve things. @vchuravy was looking at it... and from what I understand, it wasn't an enjoyable experience. If you have time, he can probably point you in the right direction.
