### 2022-12-12 #109 @devmotion
> Just use the minimal set of packages. Do you really need all the extras?

In the initial submission I loaded 5 common packages. There were zero LaTeX issues on my computer (and also compiling the LaTeX source code with Github actions worked and works flawlessly), so I did not expect them to cause any problems during the submission. Surely I could try to avoid loading any packages at all (which is not completely possible since plots are generated with pgfplots) but the main point of these packages is that they know better than I how to e.g. fix amsmath and format quotes with references correctly which IMO improves the quality of the LaTeX source code (and often also reduces its size).

To be clear, I don't want to criticize the work of the editors at all - I know how frustrating LaTeX can be and how challenging it can be to try to compile LaTeX files on a new/different computer or with a different TeXlive version. I'm just a bit frustrated that I have to deal with this online debugging here (since I can't reproduce any of the issues locally), and more generally frustrated that one still has to deal with these LaTeX package issues in 2022. Maybe something like tinytex or tectonic could be helpful? But I'm not sure if that would solve the issues with non-default packages at all. Maybe another approach would be to automatically install LaTeX packages in that are listed in some separate text file? But I guess that might open up for different security issues, so maybe that's not a good idea for a public service such as whedon.
