### 2022-12-12 #109 @odow
I only just joined the editorial team, so I've missed the backstory, but from what I understand of my one meeting: everyone agrees that things could/should be better. JOSS have updated to a GitHub action framework to build the PDFs that would give us more control. But updating the JuliaCon stuff is work that no one has the bandwidth to do.

I don't know how the bot works behind the scenes, or how to access the logs, but I wouldn't be surprised if it looks for the `paper` directory in your repo and copies that, not the entire Git repository. So that may mean it can't find your relative links like `../data` and `../src`. Could you not just copy those files into the `.tex`? The source files in particular just rely on a few lines, and they use relative references. So if you edit the text files you'll break the paper. You may as well just copy them into the tex.

> I'm just a bit frustrated that I have to deal with this online debugging here (since I can't reproduce any of the issues locally), and more generally frustrated that one still has to deal with these LaTeX package issues in 2022

Do you have a different experience submitting papers to proper Journals??? 😆😆😆
