### 2022-12-12 #109 @matbesancon
One hack that would work is add the sty packages directly in your paper, I think they should be used correctly
