### 2022-12-13 #109 @devmotion
> One hack that would work is add the sty packages directly in your paper, I think they should be used correctly

That's how I fixed the `orcidlink` issue. `mathtools` and `csquotes` I removed completely and worked around it in the LaTeX file.

It would be good to be able to inspect the logs before randomly trying more things now. Due to the different (i.e., non-existent) error message I assume the problems with unavailable LaTeX packages might be fixed and it might be a different issue now.
