### 2022-12-13 #109 @odow
> It would be good to be able to inspect the logs before randomly trying more things now

I don't know how to inspect the logs, @matbesancon or @vchuravy might know.

But it shouldn't take much trial and error.

There's a few likely candidates:
 * The use of `orchidlink` in the name: https://github.com/SciML/DelayDiffEq.jl/blob/a7cbbb7bf997114f6bb58ca18955f375542be512/paper/paper.yml#L7
 * This affiliation, https://github.com/SciML/DelayDiffEq.jl/blob/a7cbbb7bf997114f6bb58ca18955f375542be512/paper/paper.yml#L12, which should likely be `[2, 3, 4]`
 * The use of pgfplots reading from a parent directory: https://github.com/SciML/DelayDiffEq.jl/blob/a7cbbb7bf997114f6bb58ca18955f375542be512/paper/paper.tex#L17-L22 (comment out the figures, or replace them with PDFs)
 * The use of `lstlisting` from a parent directory: https://github.com/SciML/DelayDiffEq.jl/blob/a7cbbb7bf997114f6bb58ca18955f375542be512/paper/paper.tex#L497-L515 (hard-code in the lines)
