### 2022-12-24 d884c07
Remove unstable method from tests
