### 2023-02-06 #109 @devmotion
Great! And sorry, seems I forgot to reply to your comments above - I was at least certain about the first two not being problematic since I had already fixed the issues with `orcidlink` (the package was not part of the whedon installation: https://github.com/JuliaCon/proceedings-review/issues/109#issuecomment-1347332598 - adding orcidlink.sty to the repo fixed it) and the affiliations (was a ruby bug: https://github.com/JuliaCon/proceedings-review/issues/109#issuecomment-1320450434).
