+++ paper/paper.tex
@@ -15,7 +15,7 @@
 
 % configuration
 \pgfplotsset{
-    compat=1.18,
+    compat=1.16,
     % CSV files use comma as delimiter
     table/col sep=comma,
     % folders with CSV files
