### 2023-02-09 0daf7df
add a tolerance to floating point check
