### 2023-02-16 b411092
Fix test tolerance (#257)
* Fix test tolerance

* Update test/interface/fpsolve.jl

* Update test/interface/fpsolve.jl

* format
