### 2023-02-18 536be45
update to arrayinterface v7
