+++ paper/paper.tex
@@ -4,7 +4,10 @@
 \setcounter{page}{1}
 
 % math support
-\usepackage{amsmath}
+\usepackage{mathtools}
+
+% hyphenation
+\usepackage[english]{babel}
 
 % links to orcid
 \usepackage{orcidlink}
@@ -15,7 +18,7 @@
 
 % configuration
 \pgfplotsset{
-    compat=1.16,
+    compat=1.15,
     % CSV files use comma as delimiter
     table/col sep=comma,
     % folders with CSV files
@@ -23,8 +26,8 @@
     % all lines are thick
     every axis plot/.append style={thick},
     % small labels
-    legend style={font=\scriptsize},
-    ticklabel style={font=\scriptsize},
+    legend style={font=\footnotesize},
+    ticklabel style={font=\footnotesize},
     % initialize Set1-5:
     cycle list/Set1-5,
     % combine it with ’mark list*’ and 'linestyles*':
@@ -68,9 +71,10 @@
 \def\pkg#1{\texttt{#1}}
 \def\alg#1{\texttt{#1}}
 
-\begin{document}
-
+% Metadata
 \input{header}
+
+\begin{document}
 
 \maketitle
 
@@ -181,7 +185,7 @@ In Figure~\ref{fig:mackey_glass} the chaotic solution computed with \pkg{DelayDi
 \begin{figure}
     \centering
     \input{figures/mackey_glass.tex}
-    \caption{Solution of problem~\eqref{eq:mackey_glass} and its time delay embedding.}
+    \caption{Solution of problem~\eqref{eq:mackey_glass} and its time delay embedding.}%
     \label{fig:mackey_glass}
 \end{figure}
 
@@ -198,8 +202,8 @@ $x_k \colon [t_{k-1}, t_{k-1} + \tau] \to \mathbb{R}^n$ of the IVP
         x_k(t_{k-1}) &= x_{k-1}(t_{k-1}),
     \end{aligned}
 \end{equation}
-where $t_k := t_0 + k\tau$. Then the function $x$ defined by
-$x(t) := x_k(t) \Longleftrightarrow t \in [t_{k-1}, t_k)$ is a solution
+where $t_k \coloneqq t_0 + k\tau$. Then the function $x$ defined by
+$x(t) \coloneqq x_k(t) \Longleftrightarrow t \in [t_{k-1}, t_k)$ is a solution
 to the original IVP in eq.~\eqref{eq:dde_ivp}.
 
 This algorithm for solving DDEs is called method of steps. It can be applied more
@@ -297,7 +301,7 @@ Hereby we proceed in the following way:
     on time interval $[t_{k - 1}, t_k]$ by completing this discrete approximation
     using the specialized collocation polynomials in \pkg{OrdinaryDiffEq}.
     \item \label{enum:check} If in steps~\ref{enum:ode_step} and~\ref{enum:dense}
-    the "old" dense approximation was only evaluated at time points $\leq t_{k - 1}$,
+    the ``old'' dense approximation was only evaluated at time points $\leq t_{k - 1}$,
     we stop. Otherwise we repeat steps~\ref{enum:ode_step} and~\ref{enum:dense} iteratively based on the newly calculated approximations,
     until some stopping criterion is met.
 \end{enumerate}
@@ -362,7 +366,7 @@ We do not present all features and options of \pkg{DelayDiffEq} but only provide
 example for how DDEs can be declared and solved with \pkg{DelayDiffEq}. For a more
 thorough explanation we refer the reader to the extensive documentation of the
 \pkg{DifferentialEquations} framework.\footnote{The documentation is available
[…]
