+++ paper/figures/benchmark.tex
@@ -5,9 +5,13 @@
     xmode={log},
     ymode={log},
     filter discard warning={false},
-    legend pos={north east},
     width={0.95\linewidth},
-    height={1.05\linewidth},
+    height={0.8\linewidth},
+    legend style={
+        at={(axis description cs:0.5,1.05)},
+        anchor=south,
+    },
+    legend columns={3},
 ]
     \foreach \alg in {RK4,DP5,OwrenZen3,OwrenZen5,Vern6,Vern7,Vern8,Vern9,Rosenbrock23,Rodas4,Rodas5P}{
         \addplot+ table [discard if not={alg}{\alg}, x=error, y=time] {benchmark.csv};
