+++ paper/header.tex
@@ -2,8 +2,8 @@
 
 \title{DelayDiffEq: Generating Delay Differential Equation Solvers via Recursive Embedding of Ordinary Differential Equation Solvers}
 
-\author[1]{David Widmann\texorpdfstring{~\orcidlink{0000-0001-9282-053X}}{}}
-\author[2,3,4]{Chris Rackauckas\texorpdfstring{~\orcidlink{0000-0001-5850-0663}}{}}
+\author[1]{David Widmann\texorpdfstring{~\mbox{\orcidlink{0000-0001-9282-053X}}}{}}
+\author[2,3,4]{Chris Rackauckas\texorpdfstring{~\mbox{\orcidlink{0000-0001-9282-053X}}}{}}
 \affil[1]{Uppsala University, Sweden}
 \affil[2]{Massachusetts Institute of Technology, USA}
 \affil[3]{Julia Computing}
@@ -14,7 +14,7 @@
 \hypersetup{
 pdftitle = {DelayDiffEq: Generating Delay Differential Equation Solvers via Recursive Embedding of Ordinary Differential Equation Solvers},
 pdfsubject = {JuliaCon 2019 Proceedings},
-pdfauthor = {David Widmann\texorpdfstring{~\orcidlink{0000-0001-9282-053X}}{}, Chris Rackauckas\texorpdfstring{~\orcidlink{0000-0001-5850-0663}}{}},
+pdfauthor = {David Widmann\texorpdfstring{~\mbox{\orcidlink{0000-0001-9282-053X}}}{}, Chris Rackauckas\texorpdfstring{~\mbox{\orcidlink{0000-0001-9282-053X}}}{}},
 pdfkeywords = {Julia, Delay Differential Equation, Scientific Computing},
 }
 
