### 2023-03-01 #109 @devmotion
@odow I'm curious, how did you manage to install whedon locally and use it to compile the draft?

I just spent 3 hours trying to set it up locally because I still don't know what the remaining problem is - the PDF compiles fine locally and with a completely standard pdflatex GH action in our repo (as it has always done), the comments above are utterly useless, and the output of https://whedon.theoj.org is only marginally better but still undecipherable:

> Your paper failed to compile with the following errors:
>
> Looks like we failed to compile the PDF with the following error: Error producing PDF. ! LaTeX Error: Can be used only in preamble. See the LaTeX manual or LaTeX Companion for explanation. Type H for immediate help. ... l.313 \documentclass

I'm very familiar with LaTeX and happy to fix all kinds of issues but it's just not possible if I can't reproduce them and don't have access to any logs.

> But it shouldn't take much trial and error.

Well it does. As you can see from the comments above and the commit history (https://github.com/SciML/DelayDiffEq.jl/commits/paper) the only possible strategy so far is to more or less randomly try to change some things. By now I have spent almost as much time on trying to figure out what the problem with whedon is and why it fails than writing the paper. That's insane. Even if the LaTeX system can't be improved for whatever reasons and I have to use old package versions, remove packages, or include sty files (if possible, it won't be feasible for big packages) - could we just get access to proper logs? Or some clear and simple instructions for how to install and run whedon locally? Whatever I try, ruby package installation always fails, as soon as I have worked around some issue, a new one pops up.
