### 2023-03-01 #109 @devmotion
After commenting out almost everything, finally some proper error message :smile: But whedon complains about a missing font that's used and loaded in the juliacon.cls template?!
