+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DelayDiffEq"
 uuid = "bcd4f6db-9728-5f36-b5f7-82caef46ccdb"
 authors = ["Chris Rackauckas <accounts@chrisrackauckas.com>"]
-version = "5.41.0"
+version = "5.42.0"
 
 [deps]
 ArrayInterface = "4fba245c-0d91-5ea0-9b3e-6abc04ee57a9"
@@ -24,7 +24,7 @@ DiffEqBase = "6"
 OrdinaryDiffEq = "6.33"
 RecursiveArrayTools = "2"
 Reexport = "0.2, 1.0"
-SciMLBase = "1.70"
+SciMLBase = "1.89"
 SimpleNonlinearSolve = "0.1"
 UnPack = "0.1, 1.0"
 julia = "1.6"
