### 2023-03-02 #109 @devmotion
I removed the JuliaCon template and finally managed to compile the paper (in an incomplete form).
