### 2023-03-02 #109 @devmotion
Added back pgfplots (with files in subdirectory), still works.
