### 2023-03-02 #109 @devmotion
The last two runs of whedon indicate that actually the JuliaCon proceedings class is causing the issues here: https://github.com/SciML/DelayDiffEq.jl/tree/183be6029b8489f127746d4aa2ac754a2b421604 (no changes apart from documentclass article, i.e., in particular including lstinputlisting, pgfplots, and orcidlink) compiles fine with whedon but https://github.com/SciML/DelayDiffEq.jl/tree/f69f5f51d53a71610595fc93ac3b9d7c5a41fb82 (reverted to JuliaCon proceedings class, otherwise no changes to imports etc.) fails with the error message in the comment above.

@odow I'll try to figure out the problematic part of the `juliacon` class and I'm happy to open PRs regarding that template if it is clear that this is an issue of general interest. I already opened a few PRs that would arguably improve the whedon TeXLive installation (https://github.com/openjournals/heroku-buildpack-tex/pull/). I can't believe that nobody has encountered these problems with whedon + the `juliacon` before? Or maybe every other submission just includes a ton of packages, specifies fonts etc. whereas our submission only loads very few packages and does neither change fonts or layout?
