### 2023-03-02 #109 @devmotion
It seems the main culprit was the acknowledgement environment provided by `juliacon` as it was necessary to modify it to fix the whedon errors. Possibly other authors did not use the environment.
