### 2023-03-02 361b4a2
More consistent LaTeX comments
