### 2023-03-02 #109 @devmotion
Finally :tada:

@odow With some fixes of the `juliacon` class, the submission was compiled successfully by whedon. I think it is ready for review :slightly_smiling_face:
