### 2023-03-08 #125 @odow
Hmm. The links are:

 * https://juliacon.github.io/proceedings-guide/reviewer/#conflict_of_interest
 * https://juliacon.org/2022/coc/

@matbesancon is this normal?
