+++ Project.toml
@@ -24,7 +24,7 @@ DiffEqBase = "6"
 OrdinaryDiffEq = "6.33"
 RecursiveArrayTools = "2"
 Reexport = "0.2, 1.0"
-SciMLBase = "1.89"
+SciMLBase = "1.90"
 SimpleNonlinearSolve = "0.1"
 UnPack = "0.1, 1.0"
 julia = "1.6"
