+++ Project.toml
@@ -20,7 +20,7 @@ SimpleUnPack = "ce78b400-467f-4804-87d8-8f486da07d0a"
 [compat]
 ArrayInterface = "6, 7"
 DataStructures = "0.18"
-DiffEqBase = "6"
+DiffEqBase = "6.122"
 OrdinaryDiffEq = "6.33"
 RecursiveArrayTools = "2"
 Reexport = "0.2, 1.0"
