+++ Project.toml
@@ -21,7 +21,7 @@ SimpleUnPack = "ce78b400-467f-4804-87d8-8f486da07d0a"
 ArrayInterface = "6, 7"
 DataStructures = "0.18"
 DiffEqBase = "6.122"
-OrdinaryDiffEq = "6.33"
+OrdinaryDiffEq = "6.49.1"
 RecursiveArrayTools = "2"
 Reexport = "0.2, 1.0"
 SciMLBase = "1.90"
