### 2023-03-12 27bef34
Replace destats with stats (#258)
* Replace destats with stats

* Revert changes of `DDEStats` struct

* Update Project.toml

* Update Project.toml

* Update Project.toml

* only stats

---------

Co-authored-by: David Widmann <devmotion@users.noreply.github.com>
Co-authored-by: Christopher Rackauckas <accounts@chrisrackauckas.com>
