+++ test/interface/fpsolve.jl
@@ -17,10 +17,10 @@ const testsol = TestSolution(solve(prob, MethodOfSteps(Vern9());
     sol = solve(prob, alg)
 
     # check statistics
-    @test sol.destats.nf > 3000
-    @test sol.destats.nsolve == 0
-    @test sol.destats.nfpiter > 300
-    @test sol.destats.nfpconvfail > 50
+    @test sol.stats.nf > 3000
+    @test sol.stats.nsolve == 0
+    @test sol.stats.nfpiter > 300
+    @test sol.stats.nfpconvfail > 50
 
     # compare it with the test solution
     sol2 = appxtrue(sol, testsol)
@@ -31,10 +31,10 @@ const testsol = TestSolution(solve(prob, MethodOfSteps(Vern9());
     sol_oop = solve(prob_oop, alg)
 
     # compare it with the in-place solution
-    @test sol_oop.destats.nf == sol.destats.nf
-    @test sol_oop.destats.nsolve == sol.destats.nsolve
-    @test sol_oop.destats.nfpiter == sol.destats.nfpiter
-    @test sol_oop.destats.nfpconvfail == sol.destats.nfpconvfail
+    @test sol_oop.stats.nf == sol.stats.nf
+    @test sol_oop.stats.nsolve == sol.stats.nsolve
+    @test sol_oop.stats.nfpiter == sol.stats.nfpiter
+    @test sol_oop.stats.nfpconvfail == sol.stats.nfpconvfail
     @test sol_oop.t≈sol.t atol=1e-3
     @test sol_oop.u ≈ sol.u
     @test isapprox(sol.u, sol_oop.u; atol = 1e-7)
@@ -44,10 +44,10 @@ const testsol = TestSolution(solve(prob, MethodOfSteps(Vern9());
     sol_scalar = solve(prob_scalar, alg)
 
     # compare it with the in-place solution
-    @test sol_scalar.destats.nf == sol.destats.nf
-    @test sol_scalar.destats.nsolve == sol.destats.nsolve
-    @test sol_scalar.destats.nfpiter == sol.destats.nfpiter
-    @test sol_scalar.destats.nfpconvfail == sol.destats.nfpconvfail
+    @test sol_scalar.stats.nf == sol.stats.nf
+    @test sol_scalar.stats.nsolve == sol.stats.nsolve
+    @test sol_scalar.stats.nfpiter == sol.stats.nfpiter
+    @test sol_scalar.stats.nfpconvfail == sol.stats.nfpconvfail
     @test sol_scalar.t≈sol.t atol=1e-3
     @test sol_scalar.u ≈ sol[1, :]
 end
@@ -60,10 +60,10 @@ end
     sol = solve(prob, alg)
 
     # check statistics
-    @test sol.destats.nf < 2500
-    @test sol.destats.nsolve > 0
-    @test sol.destats.nfpiter < 250
-    @test sol.destats.nfpconvfail < 50
+    @test sol.stats.nf < 2500
+    @test sol.stats.nsolve > 0
+    @test sol.stats.nfpiter < 250
+    @test sol.stats.nfpconvfail < 50
 
     # compare it with the test solution
     sol2 = appxtrue(sol, testsol)
@@ -74,10 +74,10 @@ end
     sol_oop = solve(prob_oop, alg)
 
     # compare it with the in-place solution
-    @test_broken sol_oop.destats.nf == sol.destats.nf
-    @test_broken sol_oop.destats.nsolve == sol.destats.nsolve
-    @test_broken sol_oop.destats.nfpiter == sol.destats.nfpiter
[…]
