+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DelayDiffEq"
 uuid = "bcd4f6db-9728-5f36-b5f7-82caef46ccdb"
 authors = ["Chris Rackauckas <accounts@chrisrackauckas.com>"]
-version = "5.41.1"
+version = "5.42.0"
 
 [deps]
 ArrayInterface = "4fba245c-0d91-5ea0-9b3e-6abc04ee57a9"
@@ -20,11 +20,11 @@ SimpleUnPack = "ce78b400-467f-4804-87d8-8f486da07d0a"
 [compat]
 ArrayInterface = "6, 7"
 DataStructures = "0.18"
-DiffEqBase = "6"
-OrdinaryDiffEq = "6.33"
+DiffEqBase = "6.122"
+OrdinaryDiffEq = "6.49.1"
 RecursiveArrayTools = "2"
 Reexport = "0.2, 1.0"
-SciMLBase = "1.70"
+SciMLBase = "1.90"
 SimpleNonlinearSolve = "0.1"
 SimpleUnPack = "1"
 julia = "1.6"

[…]
