+++ .github/workflows/CI.yml
@@ -57,7 +57,7 @@ jobs:
           GROUP: ${{ matrix.group }}
       - uses: julia-actions/julia-processcoverage@v1
         if: matrix.version == '1' && matrix.arch == 'x64'
-      - uses: codecov/codecov-action@v2
+      - uses: codecov/codecov-action@v3
         if: matrix.version == '1' && matrix.arch == 'x64'
         with:
           file: lcov.info
