+++ .github/workflows/CI.yml
@@ -34,7 +34,7 @@ jobs:
           - version: '1.6'
             arch: x86
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.version }}
