### 2023-03-15 #125 @Datseris
Alright, I am submitting my review!

# Summary

This is an excellently written paper. It clearly introduces the type of modelling it is about, why modelling this is useful for the scientific community, and why modelling this is difficult from both a technical and a conceptual viewpoint.

The paper the proceeds in explaining clearly the brilliant design of the library and its integration with OrdinaryDiffEq.jl. This is very important. To my knowledge, OrdinaryDiffEq.jl is by far the most featureful and most performant software for solving ordinary differential equations. Integrating with that brings so many benefits into the table practically "for free". The authors have employed several "tricks" to make this integration possible, such as linking caches between the ODE solver, and the DDE solver, which in fact includes the ODE solver, or employing special handling for discontinuities, which seems to be one of the hardest technical problems in DDE solving.

All in all, this is an incredible piece of work and I would highly recommend publication!

# Code

I was able to install the software with one command. I was able to run its test suite with passing tests. The test suite was very extensive, yet one more thing I found impressive. The example tutorial was clear and straightforward, and I've also run it successfully.

# Comments
- "A disadvantage of the method of steps that it can be inefficient." needs "is" before "that".
- There are no "installation instructions" in the sense of I couldn't find anywhere the statement "To install run `Pkg.add("DelayDiffEq")`".
- It is difficult from the documentation to tell which features are part of DelayDiffEq.jl and which features are part of the greater ecosystem. Reason for this difficulty is that DelayDiffEq.jl is not independently documented, but rather documented directly in the whole DifferentialEquations.jl-wide documentation. My review checklist requires me to check that the Functionality of the software is documented, and this is the only tick I haven't checked off yet. I think I need the authors to clarify this for me. Does the page https://docs.sciml.ai/DiffEqDocs/stable/tutorials/dde_example/ contain all functionality of the software?
