### 2023-04-05 #125 @ranocha
Thanks for all the work and this manuscript. It should be published after some revisions. In addition to the comments of @Datseris above, please consider the following comments.

- [ ] Page 2, "is a scalar DDE system". Usually, people would call it a scalar DDE and not a system since it has only one component.
- [ ] Page 2, "Additionally, problems with vanishing delays cannot be solved by the method of steps.". I think this can be confusing. You write about DDEs with a fixed delay. If the delay vanishes, it's just an ODE, isn't it?
- [ ] Page 3: You talk about convergence failures. What exactly does that mean - your fixed-point iteration fails to converge? What do you do in this case?
- [ ] It would make sense to cite the articles developing the methods you mention in your manuscript.
- [ ] Page 4: It would help people not familiar with OrdinaryDIffEq.jl if you explained the parameters briefly, e.g., that `p` can be some parameters (that will just be ignored in this example).
- [ ] https://github.com/SciML/DelayDiffEq.jl/issues/265
- [ ] You have argued that the method of steps is inefficient and bad on page 2. However, you present quite a lot of examples where you use `MethodOfSteps` as algorithm. This is quite confusing. Is there an alternative?
- [ ] Page 4, "the most costly aspect for solving large stiff equations". I would argue that you just cannot factorize a Jacobian for a large system where you need thousands of MPI processes.
- [ ] Page 4, eq. (7): This is a state-dependent delay since you need something like $x_1(x_6(t))$, correct? It may be nice to point this out explicitly for readers curious about the technology but not necessarily familiar with all DDE examples.
- [ ] Reference [15] is incomplete.
- [ ] Some references such as [4] are lacking DOIs.
- [ ] It would be nice to clarify which of the arguments listed in https://docs.sciml.ai/DiffEqDocs/stable/types/dde_types/ are keyword arguments, e.g., by introducing an additional header "Keyword arguments" after the description of the "Arguments".
- [ ] It would be nice to add the same standard badges to the README of DelayDiffEq.jl as for other repos such as OrdinaryDiffEq.jl (and maybe also a link to the documentation next to the GitHub summary). Right now, there is no obvious way to get to the docs when viewing the upper part of the GitHub website of DelayDiffEq.jl.
