### 2023-04-05 SciML/DelayDiffEq.jl#265 @devmotion
Oops that's a typo: The syntax is `lag(u, p, t)` (see e.g. https://docs.sciml.ai/DiffEqDocs/stable/types/dde_types/#SciMLBase.DDEProblem).

```julia
julia> lag(u, p, t) = 1.0
lag (generic function with 2 methods)

julia> prob = DDEProblem(f, h, tspan; dependent_lags = (lag,))
DDEProblem with uType Float64 and tType Float64. In-place: false
timespan: (0.0, 10.0)
u0: 1.0

julia> sol = solve(prob, MethodOfSteps(Tsit5()))
retcode: Success
Interpolation: specialized 4th order "free" interpolation
t: 11-element Vector{Float64}:
  0.0
  1.0
  ⋮
  9.601646201268283
 10.0
u: 11-element Vector{Float64}:
 1.0
 2.220446049250313e-16
 ⋮
 0.03935344588087353
 0.019784973075655056
```
