### 2023-04-05 SciML/DelayDiffEq.jl#265 @devmotion
Fixed by https://github.com/SciML/DelayDiffEq.jl/commit/0dc1718c3f3fc4cc050e4c201438e739bde0cae9.
